make a dual package
