+++ package.json
@@ -10,13 +10,23 @@
   "private": false,
   "exports": {
     ".": {
-      "types": "./dist/index.d.cts",
-      "default": "./dist/index.cjs"
+      "import": {
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
+      },
+      "require": {
+        "types": "./dist/cjs/index.d.cts",
+        "default": "./dist/cjs/index.cjs"
+      }
     },
     "./helper": {
       "import": {
-        "types": "./dist/helper/index.d.ts",
-        "default": "./dist/helper/index.js"
+        "types": "./dist/esm/helper/index.d.ts",
+        "default": "./dist/esm/helper/index.js"
+      },
+      "require": {
+        "types": "./dist/cjs/helper/index.d.cts",
+        "default": "./dist/cjs/helper/index.cjs"
       }
     }
   },
@@ -24,7 +34,9 @@
     "e2e"
   ],
   "scripts": {
-    "build": "tsc -p tsconfig.build.json",
+    "build": "run-s -c build:*",
+    "build:esm": "tsc -p tsconfig.build.json",
+    "build:cjs": "tsup",
     "lint": "run-s -c lint:*",
     "lint:tsc": "tsc",
     "lint:eslint": "eslint .",
@@ -47,6 +59,7 @@
     "graphql": "^16.8.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.2",
+    "tsup": "^8.0.2",
     "typescript": "^5.4.2",
     "vitest": "^0.34.5",
     "vitest-github-actions-reporter": "^0.10.0"
